[FEAT] #58 - 음악 자르기 기능 및 UI 완료
