+++ Whistle/UI/Screens/Music/MusicListView.swift
@@ -99,14 +99,12 @@ struct MusicListView: View {
                   downloadAudioUsingAlamofire(for: music)
                 case .complete:
                   DispatchQueue.main.async {
+                    musicVM.musicInfo = music
                     musicVM.url = fileDirectories[music]
                     Task {
-                      if let url = musicVM.url {
-                        musicVM.sample_count = Int(audioDuration(url))
-                      }
-                      if let duration = editorVM.currentVideo?.totalDuration {
+                      if let url = musicVM.url, let duration = editorVM.currentVideo?.totalDuration {
+                        musicVM.sample_count = Int(audioDuration(url) / (duration / 10))
                         musicVM.trimmedDuration = duration
-                        print("set duration: \(duration)")
                       }
                       await musicVM.visualizeAudio()
                       bottomSheetPosition = .hidden
@@ -343,14 +341,6 @@ extension MusicListView {
   }
 }
 
-//// MARK: - MusicListView_Previews
-//
-// struct MusicListView_Previews: PreviewProvider {
-//  static var previews: some View {
-//    MusicListView(progressStatus: [:], editorVM: VideoPlayerManager)
-//  }
-// }
-
 extension UIApplication {
   func endEditing() {
     sendAction(#selector(UIResponder.resignFirstResponder), to: nil, from: nil, for: nil)
@@ -384,9 +374,6 @@ struct SearchBar: View {
             withAnimation {
               isSearching = false
             }
-          }
-          .onChange(of: searchText) { _ in
-            print("HI")
           }
         if isSearching {
           Text("취소")
