+++ Whistle/UI/Screens/TabbarView.swift
@@ -48,6 +48,7 @@ struct TabbarView: View {
       switch tabbarModel.tabSelectionNoAnimation {
       case .main:
         Color.clear
+
       case .upload:
         // FIXME: - uploadview로 교체하기
 //        Color.pink.ignoresSafeArea()]
@@ -56,6 +57,12 @@ struct TabbarView: View {
             Color.pink.ignoresSafeArea()
             PickerConfigViewControllerWrapper()
           }
+        }
+        .onAppear {
+          tabbarModel.tabbarOpacity = 0.0
+        }
+        .onDisappear {
+          tabbarModel.tabbarOpacity = 1.0
         }
       case .profile:
         NavigationStack {
@@ -71,6 +78,7 @@ struct TabbarView: View {
               .environmentObject(apiViewModel)
           }
         }
+
         .tint(.black)
       }
       VStack {
